pyproject.toml
[project]
name = "sample-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.31",
    "numpy>=2.3.2",
    "openai>=1.101.0",
    "pandas>=2.3.2",
    "pinecone>=7.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "tiktoken>=0.11.0",
]
